+++ tf_encrypted/convert/__init__.py
@@ -1,8 +1,12 @@
 """The TFE Converter."""
 from .convert import Converter
+from .inspect import inspect_subgraph, export, print_from_graphdef
 from .register import registry
 
 __all__ = [
     'Converter',
-    'registry'
+    'export',
+    'inspect_subgraph',
+    'print_from_graphdef',
+    'registry',
 ]