+++ tf_encrypted/tensor/native.py
@@ -287,7 +287,7 @@ def native_factory(NATIVE_TYPE, EXPLICIT_MODULUS=None):  # pylint: disable=inval
     def gather(self, indices: list, axis: int = 0):
       return DenseTensor(tf.gather(self.value, indices, axis=axis))
 
-    def split(self, num_split: int, axis: int = 0):
+    def split(self, num_split: Union[int, list], axis: int = 0):
       values = tf.split(self.value, num_split, axis=axis)
       return [DenseTensor(value) for value in values]